Fix/backend/participants consistency

+++ src/backend/defs/go-objects/contact.go
@@ -7,6 +7,7 @@ import (
 	"github.com/gocql/gocql"
 	"github.com/satori/go.uuid"
 	"sort"
+	"strings"
 	"sync"
 	"time"
 )
@@ -617,6 +618,7 @@ func (lookup *ContactByContactPoints) UpdateLookups(contacts ...interface{}) fun
 				oldContact := contacts[1].(*Contact)
 				for lookup := range oldContact.GetLookupKeys() {
 					lkp := lookup.(*ContactByContactPoints)
+					lkp.Value = strings.ToLower(lkp.Value)
 					lkpKey := lkp.UserID + lkp.Value + lkp.Type
 					oldLookups[lkpKey] = lkp
 					removedLookups[lkpKey] = lkp
@@ -625,6 +627,7 @@ func (lookup *ContactByContactPoints) UpdateLookups(contacts ...interface{}) fun
 			// iterate over contact's current state to add or update lookups
 			for lookup := range newContact.GetLookupKeys() {
 				lkp := lookup.(*ContactByContactPoints)
+				lkp.Value = strings.ToLower(lkp.Value)
 				lkpKey := lkp.UserID + lkp.Value + lkp.Type
 				// check if it's a new uri
 				err := session.Query(`INSERT INTO contact_lookup (user_id, value, type, contact_id) VALUES (?,?,?,?)`,
@@ -658,7 +661,6 @@ func (lookup *ContactByContactPoints) UpdateLookups(contacts ...interface{}) fun
 				// it remains lookups in the map, meaning these lookups references have been removed from contact
 				// need to cleanup lookup tables
 				for _, lookup := range removedLookups {
-					// try to get the contact_lookup
 					err := session.Query(`DELETE FROM contact_lookup WHERE user_id = ? AND value = ? AND type = ?`,
 						lookup.UserID,
 						lookup.Value,
@@ -689,22 +691,24 @@ func (lookup *ContactByContactPoints) CleanupLookups(contacts ...interface{}) fu
 	if len(contacts) == 1 {
 		contact := contacts[0].(*Contact)
 		return func(session *gocql.Session) error {
-			// seek into contact_lookup to delete references to the deleted contact
-			related, err := session.Query(`SELECT * from contact_lookup WHERE user_id = ?`, contact.UserId.String()).Iter().SliceMap()
-			if err != nil {
-				return err
-			}
-			for _, lookup := range related {
-				contact_id := lookup["contact_id"].(gocql.UUID)
-				if contact.ContactId.String() == contact_id.String() {
-					log.Info("Deleting contact_lookup for user %s with value %s", lookup["user_id"], lookup["value"])
-					err := session.Query(`DELETE FROM contact_lookup WHERE user_id = ? AND value = ? AND type = ?`,
-						lookup["user_id"],
-						lookup["value"],
-						lookup["type"]).Exec()
-					if err != nil {
-						return err
-					}
+			for lookup := range contact.GetLookupKeys() {
+				lkp := lookup.(*ContactByContactPoints)
+				lkp.Value = strings.ToLower(lkp.Value)
+				err := session.Query(`DELETE FROM contact_lookup WHERE user_id = ? AND value = ? AND type = ?`,
+					lkp.UserID,
+					lkp.Value,
+					lkp.Type,
+				).Exec()
+				if err != nil {
+					log.WithError(err).Warnf(`[CassandraBackend] UpdateLookups DELETE failed for user: %s, value: %s, type: %s`,
+						lkp.UserID,
+						lkp.Value,
+						lkp.Type)
+				}
+				// do the job of updating participants and discussion related tables
+				err = updateURIWithContact(session, lkp, false)
+				if err != nil {
+					// TODO : handle multiple error returns to merge into one error or break immediately ?
 				}
 			}
 			return nil
@@ -717,6 +721,7 @@ func (lookup *ContactByContactPoints) CleanupLookups(contacts ...interface{}) fu
 // to be able to link URI with contact when building discussions
 // if associate == true, then lkp is associated with contact, else it is dissociated from contact
 func updateURIWithContact(session *gocql.Session, lkp *ContactByContactPoints, associate bool) error {
+	lkp.Value = strings.ToLower(lkp.Value)
 	// lookup uris_hashes where lkp's uri is embedded
 	lkpUri := lkp.Type + ":" + lkp.Value
 	uriLookups, err := session.Query(`SELECT * FROM hash_lookup WHERE user_id = ? AND uri = ?`, lkp.UserID, lkpUri).Iter().SliceMap()
@@ -731,29 +736,40 @@ func updateURIWithContact(session *gocql.Session, lkp *ContactByContactPoints, a
 		var uriHash HashLookup
 		uriHash.UnmarshalCQLMap(lookup)
 		urisComponents[uriHash.Hash] = uriHash.HashComponents
-		var participants string
-		err := session.Query(`SELECT value FROM participant_hash WHERE user_id = ? AND kind = ? AND key = ?`, lkp.UserID, "uris", uriHash.Hash).Scan(&participants)
+		participants, err := session.Query(`SELECT value FROM participant_hash WHERE user_id = ? AND kind = ? AND key = ?`, lkp.UserID, "uris", uriHash.Hash).Iter().SliceMap()
 		if err != nil {
-			if err.Error() == "not found" {
-				continue
-			} else {
-				log.WithError(err).Warnf("updateURIWithContact failed to query participant_hash for user_id = %s, kind = %s, key = %s", lkp.UserID, "uris", uriHash.Hash)
-				continue
-			}
+			log.WithError(err).Warnf("updateURIWithContact failed to query participant_hash for user_id = %s, kind = %s, key = %s", lkp.UserID, "uris", uriHash.Hash)
+			continue
 		}
-		row := map[string]interface{}{}
-		err = session.Query(`SELECT * FROM participant_hash WHERE user_id = ? AND kind = ? AND key = ?`, lkp.UserID, "participants", participants).MapScan(row)
-		if err != nil {
-			if err.Error() == "not found" {
-				continue
-			} else {
-				log.WithError(err).Warnf("updateURIWithContact failed to query participant_hash for user_id = %s, kind = %s, key = %s", lkp.UserID, "participants", uriHash.Hash)
-				continue
+		if len(participants) == 0 {
+			// something is inconsistent…
+			// all participants_hash should have an entry in participant_hash table
+			// insert entry to participantHashes array, thus it'll be inserted in db again
+			participantHash := ParticipantHash{
+				UserId:     UUID(uuid.FromStringOrNil(lkp.UserID)),
+				Kind:       "participants",
+				Key:        uriHash.Hash,
+				Value:      uriHash.Hash,
+				Components: uriHash.HashComponents,
 			}
+			participantHashes = append(participantHashes, participantHash)
+			continue
 		}
-		var participantHash ParticipantHash
-		participantHash.UnmarshalCQLMap(row)
-		participantHashes = append(participantHashes, participantHash)
+		for _, p := range participants {
+			row := map[string]interface{}{}
+			err = session.Query(`SELECT * FROM participant_hash WHERE user_id = ? AND kind = ? AND key = ?`, lkp.UserID, "participants", p["value"]).MapScan(row)
+			if err != nil {
+				if err.Error() == "not found" {
+					continue
+				} else {
+					log.WithError(err).Warnf("updateURIWithContact failed to query participant_hash for user_id = %s, kind = %s, key = %s", lkp.UserID, "participants", uriHash.Hash)
+					continue
+				}
+			}
+			var participantHash ParticipantHash
+			participantHash.UnmarshalCQLMap(row)
+			participantHashes = append(participantHashes, participantHash)
+		}
 	}
 
 	// for each participant_hash :
@@ -770,7 +786,7 @@ func updateURIWithContact(session *gocql.Session, lkp *ContactByContactPoints, a
 			err = StoreURIsParticipantsBijection(
 				session,
 				UUID(uuid.FromStringOrNil(lkp.UserID)),
-				participantHash.Value,
+				newParticipantHash.Value,
 				newParticipantHash.Key,
 				urisComponents[participantHash.Value],
 				newParticipantHash.Components,
@@ -779,7 +795,7 @@ func updateURIWithContact(session *gocql.Session, lkp *ContactByContactPoints, a
 				log.WithError(err).Warn("StoreURIsParticipantsBijection failed")
 			} else {
 				// remove former bijection
-				err = RemoveURIsParticipantsBijection(session, participantHash)
+				err = RemoveURIsParticipantsBijection(session, participantHash, newParticipantHash)
 				if err != nil {
 					log.WithError(err).Warn("RemoveURIsParticipantsBijection failed")
 				}
@@ -812,29 +828,31 @@ func StoreURIsParticipantsBijection(session *gocql.Session, userId UUID, uriHash
 
 // RemoveURIsParticipantsBijection delete uris_hash <-> participants_hash bijection
 // in participant_hash table
-func RemoveURIsParticipantsBijection(session *gocql.Session, participantHash ParticipantHash) error {
-	var first, second string
-	first = participantHash.Kind
-	if first == "participants" {
-		second = "uris"
-	} else {
-		second = "participants"
-	}
-	e1 := session.Query(`DELETE FROM participant_hash WHERE user_id = ? AND kind = ? AND key = ? AND value = ?`,
-		participantHash.UserId,
-		first,
-		participantHash.Key,
-		participantHash.Value).Exec()
-	e2 := session.Query(`DELETE FROM participant_hash WHERE user_id = ? AND kind = ? AND key = ? AND value = ?`,
-		participantHash.UserId,
-		second,
-		participantHash.Value,
-		participantHash.Key).Exec()
-	switch {
-	case e1 != nil:
-		return e1
-	case e2 != nil:
-		return e2
+func RemoveURIsParticipantsBijection(session *gocql.Session, former, new ParticipantHash) error {
+	if former.Value != new.Value || former.Key != new.Key { // prevent removing the new bijection that is equal to the former one
+		var first, second string
+		first = former.Kind
+		if first == "participants" {
+			second = "uris"
+		} else {
+			second = "participants"
+		}
+		e1 := session.Query(`DELETE FROM participant_hash WHERE user_id = ? AND kind = ? AND key = ? AND value = ?`,
+			former.UserId,
+			first,
+			former.Key,
+			former.Value).Exec()
+		e2 := session.Query(`DELETE FROM participant_hash WHERE user_id = ? AND kind = ? AND key = ? AND value = ?`,
+			former.UserId,
+			second,
+			former.Value,
+			former.Key).Exec()
+		switch {
+		case e1 != nil:
+			return e1
+		case e2 != nil:
+			return e2
+		}
 	}
 	return nil
 }
